requirements.txt
annotated-types==0.6.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astroid==2.15.8
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
autopep8==2.0.4
Babel==2.13.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
comm==0.1.4
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
dnspython==2.4.2
docstring-to-markdown==0.13
email-validator==2.1.0.post1
executing==2.0.1
fastapi==0.103.2
fastjsonschema==2.18.1
flake8==6.1.0
fqdn==1.5.1
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.26.0
ipython==8.16.1
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
isort==5.12.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_client==8.5.0
jupyter_core==5.4.0
jupyter_server==2.9.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.7
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
lazy-object-proxy==1.9.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==3.0.2
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
nest-asyncio==1.5.8
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.1
overrides==7.4.0
packaging==23.2
pandas==2.1.2
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.11.0
pluggy==1.3.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psutil==5.9.6
pure-eval==0.2.2
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydocstyle==6.3.0
pyflakes==3.1.0
Pygments==2.16.1
PyJWT==2.8.0
pylint==2.17.7
pymongo==4.5.0
PyQt6==6.4.2
pyqt6-plugins==6.4.2.2.3
PyQt6-Qt6==6.4.3
PyQt6-sip==13.6.0
pyqt6-tools==6.4.2.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.8.2
python-multipart==0.0.6
pytoolconfig==1.2.6
pytz==2023.3.post1
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.1
qt6-applications==6.4.3.2.3
qt6-tools==6.4.3.1.3
qtconsole==5.4.4
QtPy==2.4.1
referencing==0.30.2
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rope==1.10.0
rpds-py==0.10.6
scikit-learn==1.3.2
scipy==1.11.3
Send2Trash==1.8.2
six==1.16.0
sklearn==0.0.post10
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.5
stack-data==0.6.3
starlette==0.27.0
terminado==0.17.1
threadpoolctl==3.2.0
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.12.1
tornado==6.3.3
traitlets==5.12.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
ujson==5.8.0
uri-template==1.3.0
urllib3==2.0.7
uvicorn==0.23.2
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
whatthepatch==1.0.5
widgetsnbextension==4.0.9
wrapt==1.15.0
yapf==0.40.2
zipp==3.17.0
